+++ src/gga_xc_xlyp.c
@@ -53,13 +53,13 @@ void gga_xc_xlyp_end(void *p_)
 }
 
 static void 
-gga_xc_xlyp(void *p_, double *rho, double *sigma,
-            double *e, double *vrho, double *vsigma)
+gga_xc_xlyp(void *p_, FLOAT *rho, FLOAT *sigma,
+            FLOAT *e, FLOAT *vrho, FLOAT *vsigma)
 {
-  static double cc[3] = {0.722, 0.347, 1.0};
+  static FLOAT cc[3] = {0.722, 0.347, 1.0};
 
   xc_gga_type *p = p_;
-  double dd, e1, vrho1[2], vsigma1[3];
+  FLOAT dd, e1, vrho1[2], vsigma1[3];
   int ifunc, is, js;
 
   dd = 1.0 - cc[0] - cc[1];
